Solves infinite loop - restores posting functionality
